Refactor cart items API endpoints
